Added talents, encounter, some interactivity
There are some talents. For now one is randomly selected at the beginning
This allows some physical / magical reduction / dodge

Encounters exist, to spend gold, and recover life

For now life / mana pot are not used (to come...)

Mana is useless! and as enemy are still OS, phys damage is useless too...